110 - Gemini File Search Tool/pyproject.toml
[project]
name = "gemini-file-search-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "numpy>=2.4.0",
    "python-dotenv>=1.2.1",
]
